programe finished

+++ P14/P14.c
@@ -12,9 +12,9 @@ Aim      : To write a C program to calculate the mean, variance and standard dev
  
  
 //function declaration
-int mean(int array[50],int n);
-int variance(int array[50],int size , float n);
-int SD(float n);
+float mean(int array[50],int n);
+float variance(int array[50],int size , float n);
+float SD(float n);
  
 //global variables;
 int i;
@@ -34,17 +34,17 @@ int main(){
     }
     //get the means
     means = mean(num,size);
-    printf("Means = %f",means);
+    printf("Means = %.3f",means);
     //get the variance
     vC = variance(num,size,means);
-    printf("\nVariance = %f",vC);
+    printf("\nVariance = %.3f",vC);
     //get the standard deviation
     sD= SD(vC);
-    printf("\n Standard Deviation = %f",sD);
+    printf("\n Standard Deviation = %.3f",sD);
 }
  
 //function for calculating mean
-int mean(int array[50],int n){
+float mean(int array[50],int n){
     float sum =0;
     for(i=0;i<n;i++){
         sum += array[i];
@@ -53,7 +53,7 @@ int mean(int array[50],int n){
     return (sum/n);
 }
 //function for calculating variance
-int variance(int array[50],int size , float n){
+float variance(int array[50],int size , float n){
     float sum=0;
     for(i=0;i<size;i++){
         sum += pow((array[i]-n),2);
@@ -63,7 +63,7 @@ int variance(int array[50],int size , float n){
 }
  
 //function for calculating Standard Deviation
-int SD(float n){
+float SD(float n){
  
     return (sqrt(n));
 }
